backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
Pillow==10.4.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi
python-dotenv==1.0.1
pydantic-settings==2.5.2
pydantic[email]==2.9.2
email-validator
numpy==1.26.4
qrcode[pil]==7.4.2
aiofiles==24.1.0
apscheduler==3.10.4
requests==2.31.0
httpx
opencv-python==4.9.0.80
mediapipe==0.10.9
